New York Taxi Fare Prediction Using R Language
New York City taxi rides paint a vibrant picture of life in the city. In order to predict fare, only
data which would be available at the beginning of a ride was used. This includes pickup and
dropoff coordinates, trip distance, start time, number of passengers. The fare of taxi ride is
function of the mileage and the duration of the ride(sum of drop charge, distance charge and
time charge). The data used in this study are all subset of New York City Taxi trip data which
contains observations on around 55 million taxi rides in New York City between 2009 and 2015.
The original data set contains features as pickup and dropoff locations, as longitude and
latitude coordinates, time and date of pickup and dropoff, fare amount, passenger count. The
data was processed to extract separate feature for year, month, day, weekday, hour from the
date and time of each ride, as well as Haversine distance from the pickup and dropoff longitude
and latitude. Linear regression model is used for predicting the fare amount.